Docs: Deno built-in OpenTelemetry support section
Explain how to use Denos built-in OpenTelemetry support,
which is available since Deno 2.2. This includes running Deno
with the `--unstable-otel` flag and setting the `OTEL_DENO=true`
environment variable.

Also, move the tip about Denos built-in OpenTelemetry support
to the beginning of the "Setting up OpenTelemetry" section for
better visibility.

[ci skip]